Porting using //eve/development/MAINLINE<-//eve/staging/2016-DANIELLE at head (1035006)
Binaries submitted in 1035042
Build time : 12 minutes 4.110000 seconds
Branch : //eve/branches/development/MAINLINE/

[git-p4: depot-paths = "//eve/branches/development/MAINLINE/carbon/src/VideoPlayer/": change = 1035014]

+++ Audio2Sink.cpp
@@ -41,7 +41,7 @@ Audio2Sink::~Audio2Sink()
 	Close();
 }
 
-void Audio2Sink::Create( intptr_t audioFunction, intptr_t audioPositionFunction, int32_t outputChannel )
+void Audio2Sink::Create( int64_t audioFunction, int64_t audioPositionFunction, int32_t outputChannel )
 {
 	m_setAudioStream = reinterpret_cast<SetAudioStream>( audioFunction );
 	m_getAudioStreamPosition = reinterpret_cast<GetAudioStreamPosition>( audioPositionFunction );
